added booleans and conditional callout rendering, moved calloutLogic into JobsRoute for now, possibly to be abstracted out in future
